Use TLongObjectHashMap instead of THashMap. (#37)
* Use TLongObjectHashMap instead of THashMap.

... in order to save memory. This locks us on using longs which i think
is fine.

* Removed unused imports.

+++ src/main/java/org/apache/tinkerpop/gremlin/tinkergraph/structure/SpecializedTinkerVertex.java
@@ -185,7 +185,7 @@ public abstract class SpecializedTinkerVertex implements Vertex {
                 edge = factory.createEdge(idValue, graph, outVertexRef, inVertexRef);
             }
             ElementHelper.attachProperties(edge, keyValues);
-            graph.edges.put(edge.id(), edge);
+            graph.edges.put((long)edge.id(), edge);
             graph.getElementsByLabel(graph.edgesByLabel, label).add(edge);
 
 //            acquireModificationLock();
@@ -311,7 +311,7 @@ public abstract class SpecializedTinkerVertex implements Vertex {
             }
         }).forEach(Edge::remove);
         TinkerHelper.removeElementIndex(this);
-        graph.vertices.remove(id());
+        graph.vertices.remove((long)id());
         graph.getElementsByLabel(graph.verticesByLabel, label()).remove(this);
 
         if (graph.ondiskOverflowEnabled) {
